move endpoint managing events to the instance controller
